pkg/apis: move some libraries around
