+++ pkg/apis/scorecard/v1alpha1/register.go
