+++ pkg/apis/scorecard/v1alpha1/types.go
