+++ pkg/apis/scorecard/v1alpha1/zz_generated.deepcopy.go
